+++ lib/shelly/cli/errors.rb
@@ -1,0 +1,5 @@
+module Shelly
+  module CLI
+    class HomeNotSetError < ArgumentError; end
+  end
+end